address Copilot comments
