+++ packages/roosterjs-content-model-plugins/lib/tableEdit/editors/features/TableMover.ts
@@ -18,13 +18,13 @@ import {
     setSelection,
 } from 'roosterjs-content-model-dom';
 import type {
-    DOMInsertPoint,
     DOMSelection,
     IEditor,
     ReadonlyContentModelTable,
     Rect,
     ShallowMutableContentModelDocument,
 } from 'roosterjs-content-model-types';
+import { getNodePositionFromEvent } from '../../../utils/getNodePositionFromEvent';
 
 const TABLE_MOVER_LENGTH = 12;
 /**
@@ -180,38 +180,6 @@ function isTableTopVisible(editor: IEditor, rect: Rect | null, contentDiv?: Node
 
 function setTableMoverCursor(editor: IEditor, state: boolean, type?: 'move' | 'copy') {
     editor?.setEditorStyle(TABLE_MOVER_STYLE_KEY, state ? 'cursor: ' + type ?? 'move' : null);
-}
-
-// Get insertion point from coordinate.
-function getNodePositionFromEvent(editor: IEditor, x: number, y: number): DOMInsertPoint | null {
-    const doc = editor.getDocument();
-    const domHelper = editor.getDOMHelper();
-
-    if (doc.caretRangeFromPoint) {
-        // Chrome, Edge, Safari, Opera
-        const range = doc.caretRangeFromPoint(x, y);
-        if (range && domHelper.isNodeInEditor(range.startContainer)) {
-            return { node: range.startContainer, offset: range.startOffset };
-        }
-    }
-
-    if ('caretPositionFromPoint' in doc) {
-        // Firefox
-        const pos = (doc as any).caretPositionFromPoint(x, y);
-        if (pos && domHelper.isNodeInEditor(pos.offsetNode)) {
-            return { node: pos.offsetNode, offset: pos.offset };
-        }
-    }
-
-    if (doc.elementFromPoint) {
-        // Fallback
-        const element = doc.elementFromPoint(x, y);
-        if (element && domHelper.isNodeInEditor(element)) {
-            return { node: element, offset: 0 };
-        }
-    }
-
-    return null;
 }
 
 /**
